backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
groq==0.29.0 